Partial fix for bug#7804
Incorporate Kyle's code-review suggestions.

a) use YPCAT_SAFE_LOG_STRING and YPMATCH_SAFE_LOG_STRING
b) In lw-ypcat and lw-ypmatch, map local error code to 1 before exiting the program on error.
c) Fix max index in ypmatch when parsing command line arguments
d) If the nick name file is not found, and if -t is not specified on the
   command line, print a warning message.
e) Fix parsing the nick name file. The alias is the first string upto whitespace; the name is the rest of the string upto a possible comment.

CR: kstemen

(r29650)