Fix bug 7671 - DDNS: Handle registration in child domains
The lw-update-dns tool was assuming that the domain is the same as the DNS zone. Sometimes the authoritative zone is actually parent domain or parent zone. To handle that case, lw-update-dns now searches for the child-most zone which is authoritative for the domain and updates that zone.

CR Sriram

(r29671)