Partial fix for bug 7532 - Improve performance
This checkin splits up the cell list/provider data from the user cache. This way, it is simpler to switch the database backend for the user cache.

This checkin also moves the user cache code up to the server layer so that it can later be used by the local provider.

CR Danilo

(r30073)