Fix build break
On 64bit Linux, allow a debug version of libcrypto to be built.

On AIX, do not pass -bexpall to the linker. This is causing the linker to define memcpy in libcrypto.so, because it came from a dependent library. Redefining memcpy causes duplicate symbols in whatever links to it. I think this is causing gcc to crash in lsass.

(r30106)