MERGE -c 30322 /trunk/Sapphire -> /branches/lwidentity-5.0 Revision: 30322 Author: glennc Date: 2008-12-19 19:09:21 -0800 (Fri, 19 Dec 2008)
Log Message:
-----------
Fix for bug 8043. Cache timeout setting in GP was updating incorrect entry in lsassd.conf file.

Someone changed the name of the setting and did not update GP.

Also removing cache-purge-timeout setting from default conf file for lsassd. This setting is not actually support yet according to Danilo, and this is confusing for customers since it does not do anything.

(r30324)