Fix bug 8035 - Solaris does not boot properly with nss_lsass
librestart is sometimes called by inetd during system startup on Solaris. If getpwnam_r returns error code 146, the bootup sequence is aborted and Solaris goes into maintenance mode.

This revision changes the error code that is returned by nss_lssass when lsassd is not running so that it looks like the user/group is simply not found.

(r30339)