librpc: Trying to fix the segv caused by consecutive calls to OpenSchannel()
calling rpc_set_auth_info() with rpc_c_authn_none is supposed to free
the auth_info on the binding but the refcount is still at 2 for some reason.
So we may be leaking a binding somewhere.  more likely is that we are
overlapping with other open handles such as teh one use to enuemrate the
trusted domains.

(r30347)