MERGE: -c 30355 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 7801 - lsassd crashes during interactive login after reboot

The size of the ppResults array was not being correctly updated. It does start out with size sCount, but AD_FilterNullEntries shrinks the list. The list is grown at the end to include the entries found through AD.

The double free crash occurred because the list was shrunk, but it was freed with the original size.

(r30356)