MERGE: -c 30426 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 8080 - assert fails when logging in on console

There was an assert in batch_build.c that made sure a \ was not in the user name, because the domain name should have already been parsed out. However, the user could do domain\\user\\ , for which case  the assert was overly aggressive.

lsassd would also segfault if no username or groupname was included after the domain: domain\\ . This is fixed by checking for null.

(r30427)