MERGE: -c 30420 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 8079 - sendmail coredumps on HP-UX

For some reason dlopen("libxnet.so", ...) does not work inside of sendmail on HP-UX, unless an absolute path is used (sendmail probably changes the library path).

After dlsym failed, we were calling the NULL function pointers, which crashed sendmail. The fix is to try to load libxnet with an absolute path, and if that fails, still not call NULL pointers.

(r30447)