MERGE: -c 30562 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 8106 - lsassd crashes if it finds a duplicate uid
Uid is a unique field in the sqlite database. When a new user comes in with the same uid as an old user, the old record is replaced in the user table. The entry in the object table is not removed (correlates name and sid). So essentially looking up a user with a duplicate uid creates a dangling record in the cache.

This fix realizes that the user record is missing, then does not reference the null pointer. Furthermore, the error is logged, then the user is looked up in AD instead of the cache.

(r30563)