MERGE: -c 30875 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 7692 - segfault in lw-eventlog-cli on RHEL21

When a program closes, the loader unloads all of the libraries in the process. Before the library is unloaded, its destructor functions (if it has any) are called. The libraries are supposed to be unloaded in the opposite order they were loaded (libraries are unloaded once nothing else depends on them).

All of the krb5 library destructors free error messages from a global lookup table. The global table is a linked list with the head stored in et_list_dynamic. libgssapi_krb5 removes its error messages from the global list. libcom_err frees the entire global table.

There must be a bug in the loader on RHEL 2.1. libcom_err's destructor is sometimes (not always) called before libgssapi_krb5's destructor. This causes libgssapi_krb5 to look at the already freed global table.

The fix is to NULL out et_list_dynamic once it is freed, so that even if libgssapi_krb5's destructor is called afterwards, will not look at the list.

(krb5: r30877)