MERGE: -c 30934 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 8169 - Failed to open LDAP connection to the DC

If the DC is overloaded, it will close connections. In wireshark this shows up as the connection being acknowledged, the reset by the DC before the client can send any data. From lsassd, this shows up as ldap_open returning NULL with errno set to ENOTCONN.

The fix is to add a sleep and retry.

CR Arlene

(lsass: r30935)