MERGE: -c 30970 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix bug 8158 - users and groups are not enumerated from linked cells

Wei contributed half of this change. Basically, we need to search all linked cells along with the primary cell. The search cookie (like a bookmark) needs to be expanded to include which cell is being search along with the ldap cookie for that cell.

CR Arlene

(lsass: r30971)