Fix bug 8217 - cannot login with international usernames
Calculating the password hash was failing for local users. The multi-byte string character count should have been used to calculate the wchar16 character count, instead of the byte count.

Furthermore, the mbstrlen function is required to calulate this. That function only works if the local is set.

CR Arlene

(lsass: r31026)