Fix bug 8110 - lsmbd will not start if pid file already exists
If lsmbd is shutdown uncleanly, it will not remove its pid file (/var/run/lsmbd.pid). When lsmbd is restarted, it will fail if another process is using the old pid.

This revision lets lsmbd examine the program that is using the old pid. If the program is not lsmbd, lsmbd will continue to start.

This change only affects the code path where:
1. A pid file already exists (probably lsmbd was shutdown uncleanly).
2. The pid in the pid file points to a valid process (probably the pid was reused).

CR Arlene

(lsmb: r31078)