Partial fix for bug 8222 - unable to authenticate on mac
libsmbutils was set as a noinst library, which means it gets statically linked to by anything that needs it. This causes a problem on mac if it gets linked twice in the same binary. Most notably, logging fails.

So instead, this library needs to be a shared library.

CR Danilo

(lsmb: r31090)