Fix for reactivated bug 8177.
Feedback on this new feature required the changes below. UI from GUI now shows results of running the lw-local-user-migrate.sh script. The script also reports progress of the processing and reports final status. Also added username field to args of script so that it can process the migration and use createhomedir to properly populate the destination if needed.

C/R: Yvo and Robo for the script

(domainjoin: r31157)