Fix bug 8222 - unable to authenticate on mac
res_init was being called once for every netlogon query that was made. This only happens if the data is not cached yet. If res_init is called more than once on mac, the function closes file descriptors that don't belong to it.

This translated into lsass not loading the ad provider after the join. That caused ad logins to fail.

The fix is to only call res_init once.

CR Danilo

(netlogon: r31248)