Fix for bug 8305.
Fix for incorrectly processed maxPwdAge value from AD. The value is usually a large negative number, and the code was ignoring the value and setting all user dwDaysTillPasswordExpires values to 0.

Solution was reviewed by Kyle when we discussed the issue.

Also fixed minor typo in get_status output. the Flag values was printing an extra 'a'.

(lsass: r31314)