Partial fix for bug 7532 - winbindd vs. lsass performance Fix for bug 8280 - ldap_timeout on mac
This checkin adds ldap connection caching to lsass. When a ldap connection to the DC or GC is closed, it is added to a free list. Next time a thread requests a connection to the same address, it will get the existing connection from the free list.

I had to add in more retry logic for ldap searches since the ldap connections are now held open indefinitely. This also fixes the problem with ldap timeouts during stress.

While doing stress testing on the new code, I noticed that if more than 5 requests are made at once, some of them start failing. To fix this I increased the lwmsg connection backlog.

I also changed the ldap connection timeout so that lsassd will more quickly detect when a domain goes offline.

CR Danilo

(lsass: r31335)