Fix bug 8285 - lsass plugin fails to load on mac after reboot
This file was not being cleared on domain join, which meant there were two entries in the database. This only causes a problem when lsass starts in offline mode.

Lsass starts in offline mode on mac after a reboot. The adstate.db cache holds domain enumeration information to let lsass start in offline mode.

Thanks goes to Glenn for investigating.

(domainjoin: r31338)