MERGE: -c 31430 /branches/lwidentity-5.1 -> /branches/lwidentity-5.0
Fix bug 8319 - parent group membership is not cleared with PAC

In some cases when a user is removed from a group, getgrnam would continue to say the user is a member of the group until the cache expires, even if the user logs in. The require membership check would also think the user is in the group.

This fix removes memberships that can be reasoned to no longer exist through PAC information.

CR Danilo

(lsass: r31433)