Re-apply lost lsmb change in trunk - rev 30420:
Remove use of HP-UX xpg socket stuff, effectively reapplying rev 30420,
which was related to bug 8079.  The xpg stuff was only necessary for
handling ancillary data over sockets.  But this is not needed in lwio
since it uses lwmsg (which does need the ancillary data functionality
and therefore uses the HP-UX xpg socket stuff).

(lwio: r31465)