Fix bug 8295 - Domain user authentication fails after leaving machine idle for 5 minutes
Lwio/lwrdr makes sure only one connection is kept to a machine at a time and all requests are multiplexed through that connection. After the connection was closed by the DC, lwio/lwrdr continued to use the socket. By marking the socket as invalid when an error occurs, it will be reconnected next time the socket is needed.

CR Danilo and Sriram

(lwio: r31508)