Re-apply lost lsmb change in trunk - revs 31078, 31089, 31094, 31103: [All related to bug 8110]
31078
-----
Fix bug 8110 - lsmbd will not start if pid file already exists

If lsmbd is shutdown uncleanly, it will not remove its pid file
(/var/run/lsmbd.pid). When lsmbd is restarted, it will fail if another process
is using the old pid.

This revision lets lsmbd examine the program that is using the old pid. If the
program is not lsmbd, lsmbd will continue to start.

This change only affects the code path where:
1. A pid file already exists (probably lsmbd was shutdown uncleanly).
2. The pid in the pid file points to a valid process (probably the pid was reused).

CR Arlene

31089
-----
Fix build break on Mac
The build break was caused by revision 31078, which was to fix bug 8110.

31094
-----
Fix build break on HPUX and FreeBSD
The build break was caused by revision 31078, which was to fix bug 8110.

31103
-----
Try to fix freebsd build break again
I was using the wrong header to get struct kinfo_proc from.

(lwio: r31511)