Saving work for new user migration tool. Does not affect build and is not a part of any current component packaging.
This tool will be an end user profile migration tool, which will run from the users new AD account or former local account. The end user only needs to know their former password or new AD password to authenticate that they own the files to be migrated. I will resume work on this and integrate the solution into the build in the 5.2 product timeline.

Glenn

(domainjoin: r31516)