MERGE: -c 31599 /branches/lwidentity-5.1 -> /trunk/Sapphire
Fix bugs 8285 and 8339 - can't authenticate

Lsass was not setting the thread impersonation token that lets lwrdrd know which credentials to use when opening a connection. The lwrdrd's default method is to use the process's default credentials cache (as it is when liblwrdr is loaded). This only worked if eventlog was started after the auth provider.

CR Arlene

(lsass: r31601)