Fix for bug 8349.
Update to ConfigureLogin.sh script to add SLED and Fedora v10 to list of systems that we show a warning on when X windows is in use for the given user session. The warning tell the user that Dbus could not be restarted and that a system reboot will be needed.

(domainjoin: r31756)