MERGE -c 31756 /trunk/Sapphire -> /branches/lwidentity-5.1
Revision: 31756
Author:   glennc
Date:     2009-02-19 13:54:34 -0800 (Thu, 19 Feb 2009)

Log Message:
-----------
Fix for bug 8349.

Update to ConfigureLogin.sh script to add SLED and Fedora v10 to list of systems that we show a warning on when X windows is in use for the given user session. The warning tell the user that Dbus could not be restarted and that a system reboot will be needed.

(domainjoin: r31772)