pvfs: Fill in PvfsLookFile().
Deal with the last path componnent in a case insensitive manner.
For example /pvfs/etc/HOSTS would open /etc/hosts on the server.

This is enough to get roaming profiles saving and loading
for now.

Also fix a syntax error when allocating the full path in
PvfsCreate{File,Dir}Create().

(lwio: r32512)