Fixing up formating for lw-eventlog-cli. Now it shows a table summary that is just like windows EventLog format.
The Event and User fields are now displayed.

In the detailed event summary output, I have added the additional Data field to the output. This is the one that shows the error codes that were encountered that explains the cause of the Failure Audit entries.

Added Data field to the import/export CSV feature also, so that we can get data into the Enterprise Database for Operations Dashboard feature work.

(eventlog: r32577)