MERGE: -c 32797 /trunk/Sapphire -> /branches/lwidentity-5.1
Fix bug 8515 - Bad console logins poison subsequent console logins

The login program is unusual because if a login attempt fails badly enough, the login program will reuse the pam handle for a new username.

Lsass maintains extra state in the pam context. It was keeping its own copy of the username in that context. So when login changed the username, lsass continued to use the old username.

The fix is to always copy the username from the pam handle, and never use the previously cached copy in the context.

CR Arlene

(lsass: r32799)