Fix bug 8519 - Unable to login via GUI (on Linux)
pam_sm_setcred was changed to return PAM_IGNORE instead of PAM_SUCCESS. This caused a regression on CentOS. On that platform, the system pam modules return an error if pam_sm_setcred is called for a domain user.

The fix is to return PAM_SUCCESS for domain users, PAM_USER_UNKNOWN for non-domain users. However, PAM_IGNORE still needs to be returned instead of PAM_SUCCESS on Solaris, so that Solaris' pam_set_cred will get called (which does work for domain users).

CR Arlene

(lsass: r32891)