MERGE: -c 33064 /trunk/Sapphire -> /branches/lwidentity-5.1
Fix for bug# 8531: LWIS: Cannot enumerate user/group enabled in Default cell (schema mode) when linked to non-default cell and joined to non-default cell. Also, fix the issue when we join to an empty OU Cell which has linked cell, we stop searching linked cell because the primary cell is empty.

(lsass: r33156)