Fix bug 7582 - lsassd fails to load AD provider when trusted domains are unavailable
It is possible to restrict the machine account from accessing specific DCs or domains it would typically have access to. Previously, if lsassd was blocked from connecting due to a logon restriction, it would fail to load the ad provider. Now it will attempt to connect to other DCs in the domain, and mark the domain as offline if none can be found.

The GC offline flag in lsass is now separate from the DC offline flag, since it is possible for all DCs in the forest root domain to be offline, but yet GCs can still be accessed. The GC offline flag is only applicable to the forest root domain. It will never be set for other domains.

The NtVer is changed is netlogon so that when searching for a GC, GCs from different domains will reply. Before, the DnsDomain ldap filter was excluding child domains when searching for a GC.

The retry logic for making ldap connections is changed in lsass. When a connection fails, instead of reaffinitizing to any DC that netlogon can CLDAP ping, a black list is used to definitively connect to a new DC. Up to 5 servers are attempted before the domain is marked offline.

Bug 8541 is also fixed. Kerberos clock skew errors are translated as network related failures so that Lsass will try to connect to other DCs. If no DC can be contacted, the domain will be marked as offline (instead of failing to load the whole auth provider).

Lwio, lwbase, lsass, and netlogon CR by Danilo
Dcerpc CR by Brian
Librpc CR by Jerry

(lwio: r33181)