MERGE: -c 33251 /trunk/Sapphire -> /branches/lwidentity-5.1
Fix bug 8589 - memory leak in lsassd found during stress
Several of the out parameters in LsaDmEngineGetDomainInfoWithNT4Name and LsaDmEngineGetDomainInfoWithObjectSid are optional. If NULL is passed in, the out parameter is not set. However, the data needs to be freed if it is not returned to the caller.

CR Wei

(lsass: r33252)