Fix bug 8532 - login reprompted when sshing as local user on AIX
With openssh 5.0p1 with the UseLogin setting turned on, the first password prompt would fail for a local user. The log would show the password as expired. Ssh would give a login prompt, then a password prompt. That second prompt would succeed:
Password:
login: <username>
Password:
$

The lsass function to check whether the password is expired was being called for the /etc/passwd local user. The lam module was returning 2, which means do not allow the login. Now the correct error code of -1 is being returned (which according to the IBM docs can mean the user does not exist).

CR Arlene

(lsass: r33330)