pvfs: Implement worker thread pool for handling IRPs.
Currently the driver dispatch routine will add the new work item
to the queue and the synchronize on a condition variable
in the IrpContext so this is not technically asynchronous.
However, once the async I/O interface is completed in the
IoMgr, it will be trivial to replace the condition variable
with a STATUS_PENDING error return and have the worker trhead
invoke the callback function instead of call pthread_cond_signal().

Wraps the LwRtlQueue from lwbase in a mutex and two condition
variables (for when and empty queue has a new item and when
a full queue has an available slot).

Currently, the work queue is of size 1 with 2 worker threads
in order to stress the thread contention on the shared data
structure.  I will determine appropriate sizing after some
performance testing.

(lwio: r33773)