Improve detection of child domains in one-way trusted forests. CR: Wei (suggested minor name/comment changes in subsequent commit)
This fixes bug 8694 which is a regression cased by the fix to bug 8588.  It
also adds configuration settings for domain manager parameters (unknown domain
cache timeout, online domain check interval).

The original problem from bug 8588 was that we did not know about one-way
trusted forest child domains.  Therefore, when presented with the SID or
domain name from such a domain, we would treat them as unknown and give up.
The fix for bug 8588 addressed this, but failed to take into account SID
history such that we got errors when SID history SIDs were present.  This new
fix simplifies the logic introduced as the fix to 8588 while addressing the
SID history case to behave as before the fix for bug 8588 was introduced
(i.e., effectively just ignore such SIDs).

The fix uses two functions, LsaDmEngineGetDomainNameWithDiscovery() and
LsaDmEngineGetDomainNameAndSidByObjectSidWithDiscovery() as wrappers for
LsaDmWrapNetLookupObjectSidByName() and
LsaDmWrapGetDomainNameAndSidByObjectSid() respectively.  The "discovery"
versions of the function do additional domain discovery instead of just
consulting the current state of the domain manager.  This is the basic idea:

LsaDmEngineGetDomainNameWithDiscovery():

- If the domain manager (DM) did not know about the domain name, look it up
  via RPC and add it if we find it.

LsaDmEngineGetDomainNameAndSidByObjectSidWithDiscovery():

- If the domain manager (DM) did not know about the domain SID, look it up via
  RPC and add it if we find it.

To handle SID history,
LsaDmEngineGetDomainNameAndSidByObjectSidWithDiscovery() must not fail with an
unexpected error (i.e., an error other than LSA_ERROR_NO_SUCH_DOMAIN).  With
the prior fix to 8588, we did not lookup the domain SID per se rather just
looking up the domain of the user's SID.  This caused us to double-add a
domain causing a LSA_ERROR_DUPLICATE_DOMAINNAME.  This fixes detects the SID
history case by correctly identifying the canonical SID of the domain.  It
also returns LSA_ERROR_NO_SUCH_DOMAIN for any SIDs that do not have a current
(i.e., non-historical) domain SID prefix.

Bacause an RPC is used to determine the validity of a domain SID when the
domain manager does not recognize the SID as a domain SID, we need to make
sure that we cache unresolvable domain SID prefixes (from SID history or
otherwise) so as to not reduce batch lookup performance to serial lookup
performance.  To achieve this, an "unknown domain SID" cache is added to the
domain manager.  This cache has a relatively short lifetime (default 1 hour).
With this case, the worst performance we could have is to do N lookups every T
seconds where N is the number of unknown domain SID prefixes (i.e., the number
of domains imported via SID history) and T is the "unknown domain SID cache"
timeout.

(lsass: r33911)