Bug 8533 - *-[Mac 10.5 PPC - UT Health,NBC,MDA] Domain Join GUI fails to 2008 & 2003 DC
Include cred cache type in the string name. Domain Join GUI links to the Carbon
framework, which pulls in Mac Kerberos. Without the cred cache type, the
default for Mac's Kerberos is API: not FILE:, which is where the credentials
was written. Explicitly including FILE: type in the name insures credential will
be written to where they are intended to be stored.

(lsass: r34000)