Bug 8516 - lsassd & netlogon '--logfile' ignored when '--start-as-daemon' is specified
Changed freopen(logFile, stdout) approach for logging to the specified --logfile logFile
to fopen(logFile, "w"). Logging stdout in this manner was not correct in netlogond,
as ERROR and WARNING messages were written to stderr and not sent to the logfile.
There was also initialization issues when daemonizing, as stdin/stdout/stderr were
reopened to point to /dev/null.
CR: aberry

(lwio: r34132)