MERGE: -c 34295 /trunk/Sapphire -> /branches/lwidentity-5.2
Bug 8774-When specifying wrong OU on the Mac, subsequent join hangs

pthread_cancel() was not functioning as expected on OSX. Changed
implementation to do a read-side shutdown on the socket in use
by the thread being stopped. This causes the select() in
SMBSocketReaderMain() to return so the shutdown flag can be tested.
CR: aberry
Note: Hand merged from revision 34254, as was checked into 5.1rc initially.

(lwio: r34302)