MERGE: -c 34254 /branches/lwidentity-5.1rc -> /branches/lwidentity-5.1
Bug 8774-When specifying wrong OU on the Mac, subsequent join hangs

pthread_cancel() was not functioning as expected on OSX. Changed
implementation to do a read-side shutdown on the socket in use
by the thread being stopped. This causes the select() in
SMBSocketReaderMain() to return so the shutdown flag can be tested.
CR: bkoropoff

(lwrdr: r34439)