Add a batch write function to eventlog, and use it in the syslog reaper
Previously, only a single event could be written per call to evenlogd. That required a separate sqlite transaction for each event.

This changes introduces a batch write function where multiple events can be written with one function call. For one, this makes it easy to use precompiled statements in sqlite. Second, it allows multiple events to be written in one transaction. Since the fixed cost for starting and ending a transaction is significantly slower than the variable cost of writing an event in a transaction, this speeds up writes.

Using batched writes, 900 events can be written per second. Using individual writes, only 35 events can be written per second.

This change removes the single write function and replaces it with a batch write function. The single write function in the client library is maintained as a wrapper to the multi-write RPC call. Per Krishna and Manny, it is okay to break RPC backwards compatibility like this.

(eventlog: r34612)