Maintain next event log id when clearing event log database
Previously the clear operation in eventlogd would drop the event table and recreate it. Dropping the table resets the next record id.

The event forwarder requires the event ID to be increasing so that it can track which events have already been forwarded. This revision changes the clear function to delete all records from the event log table instead of dropping it.

This should not noticeably affect performance because the clear operation is must be performed by an admin and is rarely used.

(eventlog: r34644)