lsass: BUG 8782 - Always return SUCCESS for LocalCheckUserInList().
The LocalCheckUserInList() entry point for the require-membership-of
option was returning LSA_ERROR_NOT_HANDLED.  Since this option is
not supported by the local provider, the function should return
LSA_ERROR_SUCCESS else local logins will be disabled.  It only
worked previously when joined to a domain because the AD provider
was returning success by default when no "require-membership-of"
list was defined.

(lsass: r34868)