BUG 8972: unable to access home directory as console login on Solaris8 sparc
Added LsaOpenSession()/LsaCloseSession() to the pam_sm_authenticate() phase for
Solaris only, to force creation of the user's home directory. Needed for DTLOGIN
to work, as there is an explicit test for the existence of the home directory
before pam_sm_open_session() is called, which is where this operation normally
occurs.

CR: bkoropoff

(lsass: r35191)