pvfs: Finish the OPLOCK_BREAK_IN_PROGRESS support
When an oplock break is in progress, defer the BreakIfLocked()
call until after the OplockBreak has been acknowledged.  This
allows for example a FILE_OPEN to break a Batch oplock to
a level2.  And then an incoming FILE_OVERWRITE to break the
level2 which is not granted until the BreakAck is received.

Major changes:

* Convert the global work queue to use a work context not
  directly tied to an IrpContext
* Move the processing of operations postponed due to an oplock
  break to the global work queue (after being moved from a
  pending queue to the FCB's ready queue)
* Store a BreakInProgreress indictaor on the FCB (as well as the
  CCB) and return OPLOCK_BREAK_IN_PROGRESS from PvfsOplockBreakIfLocked()
  in such cases.  The caller then postpones the oplock break test
  using the FCB's normal PendingOps queue.

Minor changes:

* Don't test for oplock breaks on create when the caller only
  asked for FILE_READ_ATTRIBUTES, FILE_WRITE_ATTRIBUTES, or
  SYNCHRONIZE access

(lwio: r35556)