MERGE: -c 35686 /trunk/Sapphire -> /branches/lwidentity-5.2
Fix bug 8963 - domainjoin fails to update files through symlinks on read only file system

In some cases, the /etc mount point is read-only. The files on /etc/ may still be writable because they are symlinks to a read-write mount point. Previously, domainjoin would try to write to a temporary file, then move it in place of the symlink.

Now domainjoin will follow the symlinks. The backup files will be created in the directory of the real file (discovered by following all symlinks). The intermediate file will be moved into place for the real file.

(domainjoin: r35707)