Finish enforcing a cap on the ram used by the in-memory cache.
The in-memory cache prioritizes users who have logged in (and the groups they belong to), over users who have not logged in (info was only queried). However, after 30 days, logged in user data may be removed by a 0 second old object. After 60 days, logged in user data may be removed by a 5 second old object.

Furthermore, the last 10 users to log in (plus their groups) are pinned such that they are the very last entries to be removed.

(lsass: r35869)