srvsvc_winreg: Work around MIDL extension to DCE/RPC for RegQueryValue.
MIDL allows [unique] pointers for size_is() and length_is() attributes
on conformant arrays.  The DCE/RPC spec does not allow this so
as a temporary workaround, expose the [unique] pointers as a pair of [ref]
pointers (ont for the referent ID and one for the value).

With this fix it is possible to connect to lsassd's lsa/samr interfaces
using the Local Users & Groups mmc plugin from a WinXP client.

(srvsvc: r36389)