Fix bug 8954 - .k5login creation failure blocks login
At login, the lsass daemon creates a .k5login file in the user's home directory. If the home directory is a mounted NFS share with root squash enabled, then creating the file failes with EACCES.

This revision changes lsass to ignore the EACCES error when trying to create the file. If the file cannot be created, only SSO will be affected.

CR Danilo

(lsass: r36610)