MERGE: -c 36637 -> /trunk/Sapphire
Fix bug 9228 - SID to Name lookup fails for domain accounts in sam.db

When a domain user is added to a local group, a stub entry for the domain user is added in sam.db. Previously, the local provider attempted to return the stub entry if it was queried through lw-find-by-sid. Now the local provider will say the object does not exist if it is a stub.

CR Danilo

(lsass: r36639)