pvfs: Fix a deadlock between PvfsOplockBreakIfLocked & PvfsProcessPendingLocks
When break an oplock, we free the oplock record which will in turn
release the CCB stored there.  If the release results in freeing the last
reference to the CCB, we will call PvfsUnlockFile which initiates the
processing of any pending locks since the lock table has potentially
changed.

Originally the pending BRL processing and the Oplock break processing
used the same mutex (FCB.ControlBlock) resulting in the deadlock.
The solution is to assign individual mutexes/rwLocks to BRL data
structures and to Oplock information.

(lwio: r36962)