srv-smbv1: Dont' send a BreakAck to the file system on a level2 break.
Sorry Sriram.  I told  you wrong on this.  After reviewing
the pvfs code, I remembered that a level2 break will not
block the operation that caused the break.  So there is no
STATUS_PENDING and hence no need to send an ack to initiate
processing of deferred ops.

(lwio: r37015)