srv-smbv1: Close the window between breaking an oplock and processing it
I believe this fixes this last known oplock issue.  The gist is that
there was a window between when an oplock was broken internally and
when it was processed by SRV.  During this time, the client could
close the file.  In SrvProcessOplock() the lookup for the file would
fail and therefore no BreakAck (if necessary) would be issued to the
file system.  Without an Ack, all pending calls waiting on the
acknowledgement would block forever.

This change is based on the fact that an oplock is either (a) active
(and therefore cancellable), or (b) broken.  A broken oplock is either
not been processed yet or awaiting an ack from the client, but both
cases can be handled the same in the close rundown.

So during close, we blindy try to cancel the oplock and simply
manually do the rundown if the IoCancel fails assuming that the oplock
has already been broken.

(lwio: r37061)