Fix bug 9260 - lsass crash during stress
During an ldap search, lsass received back a NULL ldap message. This appears to be due to bad error code propogation inside of openldap when it receives a nod (notice of disconnect).

All of the ldap traffic is sealed (instead of just signed), so I cannot see exactly what the DC is sending for the nod. However, I suspsect that the sealing is actually causing the nod.

This revision restores the ability to turn on and off sealing from lsassd.conf. Revision 12376 was lost when we switched to kerberos 1.7. I am restoring that change with a comment so that it is less likely to be lost in the future.

Here is the original log message:
r12376 | krishnag | 2007-01-03 15:16:15 -0800 (Wed, 03 Jan 2007) | 2 lines

Force MIT gss krb implementation to not explicitly force sign and seal

(krb5: r37241)