MERGE: -c 37255 /trunk/Sapphire -> /branches/lwidentity-5.3
Fix bug 9337 - bad password for domainjoin shows bad error message

When domainjoin was passed a bad password, the error message was:
0x9C6F - The call to Kerberos 5 failed

This is regression from moving to kerberos 1.7. In 1.7, when it receives a preauth failed message from the KDC, it retries instead of directly returning the error. Eventually kerberos fails with a too many retries error.

This change checks for a very specific kind of preauth failed message. Now when kerberos receives the message, it will bail out immediately.

The better error code generates a better error message.

CR Arlene

(krb5: r37256)