Fix bug 9399 - gpagentd segfaults under authstress
This checkin is on behalf of Brian.

The glibc nsswitch interface on Linux allows multithreaded programs to call into the lsass nsswitch library on separate threads. A shared connection is used to lsass, and that is stored in a global variable.

When gpagentd calls the unthread safe nsswitch library from multiple threads, gpagentd segfaults. The fix is to make the nsswitch library safe with a mutex around the global lsassd connection.

In general it is not safe to link an nsswitch library with libpthread, because that causes libpthread to get loaded in calling libraries at runtime, which makes them suddenly multithreaded. However on at least modern versions of Linux, pthread_mutex_lock is available in libc. This revision uses a global lock only if it is safe to do so without linking against libpthread.

CR Kyle

(lsass: r37569)