Additional fix for bug 9154 - improve logging
This revision changes the log message for AD_NetLookupObjectNameBySid so that it prints the sid that failed. I also changed the batching code to print a message when a user cannot be found by sid because it is disabled, and a symbolic error code name is used for ldap connection errors.

(lsass: r37733)