iomgr: Add IO_OPEN_FILE_INFO_100 which includes delete information.
Use level 100 rather than 1 as this information tends to be specifically
held by the filesystem.  The information is reachable via a network
protocol driver using IoQueryInformationFile() but I expect that
it too heavy weight to be a real use.

The idea is to devide open file info levels as follows:

* 0 - 99: general purpose information levels possibly supported
  by all drivers
* 100 - 199: File system specific information levels
* 200 - 299: Network protocol driver info levels.

(lwio: r37801)