pvfs: Don't remove cancelled oplocks during break processing.
Cancelled records are removed during PvfsOplockCleanOplockQueue().
This fixes a CCB ref count bug (which was in turn creating a DeleteOnClose
failure) because the oplock record could not be found during the
cancellation cleanup work item.  The end result is that the Irp
was never completed, the IrpContext was never freed, and the OplockRecord
was not being properly freed.

(lwio: r37810)