Fix build break due to bad typing on my part moving this change from my 5.3 checkout to Sapphire.
This corrects the following checkin...

Revision: 37982
Author:   glennc
Date:     2009-10-13 16:17:54 -0700 (Tue, 13 Oct 2009)

Log Message:
-----------
Fix for Shell and Homedir paths when User object Cell settings are blank. The "" string is valid and should cause lsass to use the Unprovisioned mode solution for assigning these values to the user. This would make it possible for Mac, Unix, Linux systems to default to the OS default values which are more appropriate.

(lsass: r37997)