pvfs: Fix a logic error in the CCB->OplockState state machine.
PvfsOplockMarkPendedOpsReady() was resetting the handle's OplockState
to NONE.  However, when being called form PvfsOplockBreakAck()
we could have granted the handle a level2 oplock.  Resetting the state
to NONE again causes an assert to fire when later breaking the level2
oplock.

Remove pCcb from PvfsOplockMarkPendedOpsReady() since it is no longer
needed in the function.  The assumption is that this function will
only be called when the handle in the BREAK_IN_PROGRESS state is
being cleared.

(lwio: r38226)