Simplify how principal names are passed to GSS for ldap connections.
We were redeclaring a private KRB5 oid in lwadvapi. That oid means the username is a preparsed krb5_principal. We had to call kerberos to parse the UPN into that format.

Now we are using a public KRB5 oid to pass the UPN in string format. Kerberos still does the parsing internally.

I also changed the mech oid to the public variable.

(lwadvapi: r38397)