Updating CredentialsDialog to add missing dialog interaction code to fill in textboxes, act on Okay/Cancel button actions etc.
I am using this dialog in the FileBrowser LMC plugin when credentials are needed to access a remote network share.

Relabled some of the radio button items and textbox names to match there purpose.

Added code to gray out the user creds text fields when the "Use Logged on User creds" radio box is selected. Ungray the fields when they are needed. etc.

(lmc: r38446)